Fixed a few complaints from g++ using '-pedantic'.

+++ include/elemental/advanced/ApplyRowPivots.hpp
@@ -181,7 +181,7 @@ ApplyRowPivots
     const int b = image.size();
 #ifndef RELEASE
     PushCallStack("ApplyRowPivots");
-    if( A.Height() < b || b != preimage.size() )
+    if( A.Height() < b || b != (int)preimage.size() )
         throw std::logic_error
         ("image and preimage must be vectors of equal length that are not "
          "taller than A.");
